refactor(reward): split streams.rs into modules
