+++ actors/reward/src/streams/tests/queue.rs
@@ -1,145 +1,24 @@
-use fvm_ipld_encoding::RawBytes;
+use fil_actors_runtime::BURNT_FUNDS_ACTOR_ADDR;
 use fvm_shared::address::Address;
-use fvm_shared::bigint::BigInt;
-use fvm_shared::clock::EPOCH_UNDEFINED;
+use fvm_shared::clock::{ChainEpoch, EPOCH_UNDEFINED};
 use fvm_shared::econ::TokenAmount;
 use hex_literal::hex;
-use num_traits::{ToPrimitive, Zero};
+use num_traits::Zero;
 
 use super::*;
-
-fn pct(value: u64) -> u64 {
-    DENOM / 100 * value
-}
-
-fn constant_weight(value: u64) -> WeightRecord {
-    WeightRecord { v_start: value, slope: 0, t_start: 0, floor: value, cap: value }
-}
-
-fn shares(rows: &[(u64, u64)]) -> Vec<RecipientShare> {
-    rows.iter()
-        .map(|(id, share)| RecipientShare { recipient: Address::new_id(*id), share: *share })
-        .collect()
-}
-
-fn full_share_map(first_recipient: u64) -> Vec<RecipientShare> {
-    let share = DENOM / MAX_RECIPIENTS as u64;
-    (0..MAX_RECIPIENTS)
-        .map(|idx| RecipientShare {
-            recipient: Address::new_id(first_recipient + idx as u64),
-            share,
-        })
-        .collect()
-}
-
-fn explicit(writer: u64, shares: Vec<RecipientShare>) -> ExplicitDistribution {
-    ExplicitDistribution {
-        writer: Address::new_id(writer),
-        shares,
-        payable: Vec::new(),
-        claimed_period: Vec::new(),
-    }
-}
-
-fn stream(id: StreamId, weight: u64, distribution: Option<ExplicitDistribution>) -> Stream {
-    Stream { id, weight: constant_weight(weight), distribution }
-}
-
-fn base_state() -> (StreamsState, Vec<StreamAccrual>) {
-    (
-        StreamsState {
-            streams: vec![
-                stream(1, pct(60), None),
-                stream(2, pct(20), Some(explicit(200, shares(&[(101, DENOM)])))),
-            ],
-            tombstones: Vec::new(),
-            pending_writes: Vec::new(),
-        },
-        vec![StreamAccrual { id: 2, amount: TokenAmount::zero() }],
-    )
-}
+use crate::streams::distribution::recipient_union_len;
+use crate::streams::queue::{
+    cancel_pending, project_due_writes, remove_stream, replace_writer, validate_pending_queue,
+};
+use crate::streams::weights::{WeightRecordsPayload, compute_weight};
 
 fn next_epoch(streams: &StreamsState) -> ChainEpoch {
     streams.pending_writes.first().map_or(EPOCH_UNDEFINED, |write| write.effective_epoch)
 }
 
-fn tombstone(id: StreamId, first_recipient: u64, rows: usize) -> Tombstone {
-    Tombstone {
-        id,
-        payable: (0..rows)
-            .map(|offset| RecipientAmount {
-                recipient: Address::new_id(first_recipient + offset as u64),
-                amount: TokenAmount::from_atto(1),
-            })
-            .collect(),
-    }
-}
-
-fn delegated_address() -> Address {
-    Address::new_delegated(10, &[1; 20]).unwrap()
-}
-
-#[test]
-fn rejects_non_id_addresses_in_persisted_state_and_pending_payloads() {
-    let (base, accruals) = base_state();
-    let assert_invalid = |streams: StreamsState| {
-        let error = validate_streams_state(&streams, &accruals, 0).unwrap_err();
-        assert!(error.to_string().contains("not an ID address"), "{error}");
-    };
-
-    let mut streams = base.clone();
-    streams.streams[1].distribution.as_mut().unwrap().writer = delegated_address();
-    assert_invalid(streams);
-
-    let mut streams = base.clone();
-    streams.streams[1].distribution.as_mut().unwrap().shares[0].recipient = delegated_address();
-    assert_invalid(streams);
-
-    let mut streams = base.clone();
-    streams.streams[1].distribution.as_mut().unwrap().payable =
-        vec![RecipientAmount { recipient: delegated_address(), amount: TokenAmount::from_atto(1) }];
-    assert_invalid(streams);
-
-    let mut streams = base.clone();
-    streams.streams[1].distribution.as_mut().unwrap().claimed_period =
-        vec![RecipientAmount { recipient: delegated_address(), amount: TokenAmount::from_atto(1) }];
-    assert_invalid(streams);
-
-    let mut streams = base.clone();
-    streams.tombstones = vec![Tombstone {
-        id: 3,
-        payable: vec![RecipientAmount {
-            recipient: delegated_address(),
-            amount: TokenAmount::from_atto(1),
-        }],
-    }];
-    assert_invalid(streams);
-
-    let mut streams = base.clone();
-    streams.pending_writes = vec![PendingWrite {
-        id: Some(3),
-        op: PendingWriteOp::RegisterStream,
-        payload: RawBytes::serialize(&RegisterStreamPayload {
-            weight: constant_weight(pct(10)),
-            distribution: Some(DistributionInit {
-                writer: delegated_address(),
-                shares: shares(&[(103, DENOM)]),
-            }),
-        })
-        .unwrap(),
-        effective_epoch: 10,
-    }];
-    assert_invalid(streams);
-
-    let mut streams = base;
-    streams.pending_writes = vec![PendingWrite {
-        id: Some(2),
-        op: PendingWriteOp::SetDistribution,
-        payload: RawBytes::serialize(&SetDistributionPayload { writer: delegated_address() })
-            .unwrap(),
-        effective_epoch: 10,
-    }];
-    assert_invalid(streams);
+fn random_u64(state: &mut u64) -> u64 {
+    *state = state.wrapping_mul(6_364_136_223_846_793_005).wrapping_add(1);
+    *state
 }
 
 #[test]
@@ -206,6 +85,7 @@ fn deferred_payloads_have_stable_cbor() {
         .bytes()
     );
 }
+
 #[test]
 fn removal_admission_reserves_tombstone_rows_and_claim_relieves_the_bound() {
     let (base, accruals) = base_state();
@@ -222,14 +102,6 @@ fn removal_admission_reserves_tombstone_rows_and_claim_relieves_the_bound() {
     let paid = claim(&mut over, &accruals, 3, &[Address::new_id(1_000)]).unwrap();
     assert_eq!(vec![TokenAmount::from_atto(1)], paid);
     queue_remove_stream(&mut over, 0, 1, 2).unwrap();
-}
-
-#[test]
-fn rejects_persisted_tombstone_rows_above_the_bound() {
-    let (mut streams, accruals) = base_state();
-    streams.tombstones = vec![tombstone(3, 1_000, MAX_TOMBSTONE_ROWS + 1)];
-    let error = validate_streams_state(&streams, &accruals, 0).unwrap_err();
-    assert!(error.to_string().contains("tombstone row reservation 257"), "{error}");
 }
 
 #[test]
@@ -255,280 +127,6 @@ fn pending_removal_bound_survives_share_folds_during_the_hold() {
     apply_due_writes(&mut streams, &mut accruals, 1).unwrap();
     let rows: usize = streams.tombstones.iter().map(|row| row.payable.len()).sum();
     assert!(rows <= MAX_TOMBSTONE_ROWS);
-}
-
-fn amount(rows: &[RecipientAmount], recipient: u64) -> TokenAmount {
-    rows.iter()
-        .find(|row| row.recipient == Address::new_id(recipient))
-        .map_or_else(TokenAmount::zero, |row| row.amount.clone())
-}
-
-fn service_liabilities(streams: &StreamsState, accruals: &[StreamAccrual]) -> TokenAmount {
-    compute_service_liability(streams, accruals).unwrap()
-}
-
-fn assert_service_conserved(
-    gross: &TokenAmount,
-    paid: &TokenAmount,
-    burned: &TokenAmount,
-    streams: &StreamsState,
-    accruals: &[StreamAccrual],
-) {
-    let mut accounted = paid.clone();
-    accounted += burned;
-    accounted += service_liabilities(streams, accruals);
-    assert_eq!(*gross, accounted);
-}
-
-#[derive(Default)]
-struct SupplyTracker {
-    total_minted: TokenAmount,
-    total_burn: TokenAmount,
-    total_service: TokenAmount,
-    total_dust: TokenAmount,
-    f099_balance: TokenAmount,
-    actor_balance: TokenAmount,
-}
-
-impl SupplyTracker {
-    fn award(
-        &mut self,
-        streams: &StreamsState,
-        accruals: &mut [StreamAccrual],
-        epoch: ChainEpoch,
-        reward: TokenAmount,
-    ) {
-        let allocation = allocate_reward(&streams.streams, epoch, &reward).unwrap();
-        assert!(allocation.schedule_valid, "valid randomized state entered degradation");
-        let service =
-            allocation.service.iter().fold(TokenAmount::zero(), |total, row| total + &row.amount);
-        accrue_service(accruals, &allocation.service).unwrap();
-        self.total_minted += &reward;
-        self.total_burn += &allocation.burn;
-        self.total_service += &service;
-        self.f099_balance += allocation.burn;
-        self.actor_balance += service;
-    }
-
-    fn burn_dust(&mut self, dust: TokenAmount) {
-        self.actor_balance -= &dust;
-        self.total_dust += &dust;
-        self.f099_balance += dust;
-    }
-
-    fn pay_claim(&mut self, result: &[TokenAmount]) {
-        let paid = result.iter().fold(TokenAmount::zero(), |total, amount| total + amount);
-        self.actor_balance -= paid;
-    }
-
-    fn assert_invariants(&self, streams: &StreamsState, accruals: &[StreamAccrual]) {
-        let liabilities = service_liabilities(streams, accruals);
-        assert_eq!(self.actor_balance, liabilities);
-        assert!(
-            liabilities <= self.total_service,
-            "conservative service reserve is below exact liability"
-        );
-        let miner = &self.total_minted - &self.total_burn - &self.total_service;
-        assert!(miner >= TokenAmount::zero());
-        assert_eq!(self.f099_balance, &self.total_burn + &self.total_dust);
-    }
-}
-
-fn random_u64(state: &mut u64) -> u64 {
-    *state = state.wrapping_mul(6_364_136_223_846_793_005).wrapping_add(1);
-    *state
-}
-
-fn compute_weight_unbounded(record: &WeightRecord, epoch: ChainEpoch) -> u64 {
-    let delta = BigInt::from(epoch) - BigInt::from(record.t_start);
-    let value = BigInt::from(record.v_start) + BigInt::from(record.slope) * delta;
-    value.min(BigInt::from(record.cap)).max(BigInt::from(record.floor)).to_u64().unwrap()
-}
-
-#[test]
-fn computes_weights_without_overflow_at_integer_boundaries() {
-    let values = [0, 1, DENOM, u64::MAX];
-    let slopes = [i64::MIN, i64::MIN + 1, -1, 0, 1, i64::MAX];
-    let epochs = [i64::MIN, -1, 0, 1, i64::MAX];
-    let bounds = [0, DENOM, u64::MAX];
-
-    for &v_start in &values {
-        for &slope in &slopes {
-            for &t_start in &epochs {
-                for &epoch in &epochs {
-                    for &floor in &bounds {
-                        for &cap in &bounds {
-                            let record = WeightRecord { v_start, slope, t_start, floor, cap };
-                            assert_eq!(
-                                compute_weight_unbounded(&record, epoch),
-                                compute_weight(&record, epoch),
-                                "record {record:?} at epoch {epoch}"
-                            );
-                        }
-                    }
-                }
-            }
-        }
-    }
-}
-
-#[test]
-fn computes_signed_clamped_weights() {
-    let rising = WeightRecord { v_start: 10, slope: 2, t_start: 5, floor: 4, cap: 20 };
-    assert_eq!(4, compute_weight(&rising, 0));
-    assert_eq!(10, compute_weight(&rising, 5));
-    assert_eq!(20, compute_weight(&rising, 10));
-
-    let falling = WeightRecord { v_start: 10, slope: -2, t_start: 5, floor: 4, cap: 20 };
-    assert_eq!(20, compute_weight(&falling, 0));
-    assert_eq!(10, compute_weight(&falling, 5));
-    assert_eq!(4, compute_weight(&falling, 10));
-
-    let flat = WeightRecord { v_start: 20, slope: 0, t_start: 0, floor: 4, cap: 20 };
-    assert_eq!(20, compute_weight(&flat, i64::MIN));
-    assert_eq!(20, compute_weight(&flat, i64::MAX));
-
-    let extreme =
-        WeightRecord { v_start: DENOM, slope: i64::MIN, t_start: i64::MAX, floor: 0, cap: DENOM };
-    assert_eq!(DENOM, compute_weight(&extreme, i64::MIN));
-
-    let overflowing = WeightRecord {
-        v_start: u64::MAX,
-        slope: i64::MIN,
-        t_start: i64::MAX,
-        floor: 0,
-        cap: u64::MAX,
-    };
-    assert_eq!(u64::MAX, compute_weight(&overflowing, i64::MIN));
-
-    let reversed = WeightRecord { v_start: 10, slope: 0, t_start: 0, floor: 20, cap: 4 };
-    assert_eq!(20, compute_weight(&reversed, 0));
-}
-
-#[test]
-fn evaluates_a_delayed_canonical_ramp_flat_then_rising() {
-    let delayed = WeightRecord { v_start: 10, slope: 2, t_start: 5, floor: 10, cap: 20 };
-    validate_weight_record(&delayed).unwrap();
-    assert_eq!(10, compute_weight(&delayed, 0));
-    assert_eq!(10, compute_weight(&delayed, 4));
-    assert_eq!(10, compute_weight(&delayed, 5));
-    assert_eq!(12, compute_weight(&delayed, 6));
-    assert_eq!(20, compute_weight(&delayed, 10));
-}
-
-#[test]
-fn finds_clamp_transition_edges_at_large_epochs() {
-    let record = WeightRecord { v_start: 10, slope: 3, t_start: 100, floor: 4, cap: 20 };
-    let epochs = weight_breakpoints(&record, 0);
-    for epoch in [0, 97, 98, 99, 100, 102, 103, 104, 105] {
-        assert!(epochs.contains(&epoch), "missing breakpoint {epoch}");
-    }
-
-    let record =
-        WeightRecord { v_start: 10, slope: 1, t_start: i64::MAX - 1, floor: 0, cap: DENOM };
-    let epochs = weight_breakpoints(&record, i64::MAX - 2);
-    assert!(epochs.contains(&i64::MAX));
-}
-
-#[test]
-fn validates_weight_records_and_piecewise_schedule() {
-    assert!(
-        validate_weight_record(&WeightRecord {
-            v_start: 0,
-            slope: 0,
-            t_start: 0,
-            floor: 2,
-            cap: 1,
-        })
-        .is_err()
-    );
-    assert!(validate_weight_record(&constant_weight(DENOM + 1)).is_err());
-
-    let valid = vec![
-        stream(1, pct(60), None),
-        stream(2, pct(40), Some(explicit(200, shares(&[(101, DENOM)])))),
-    ];
-    validate_weight_schedule(&valid, 0).unwrap();
-
-    let invalid = vec![
-        stream(1, pct(60), None),
-        stream(2, pct(41), Some(explicit(200, shares(&[(101, DENOM)])))),
-    ];
-    let error = validate_weight_schedule(&invalid, 0).unwrap_err();
-    assert!(error.to_string().contains("exceed DENOM"));
-
-    let crossing = vec![
-        Stream {
-            id: 1,
-            weight: WeightRecord {
-                v_start: pct(50),
-                slope: 1,
-                t_start: 0,
-                floor: pct(50),
-                cap: pct(70),
-            },
-            distribution: None,
-        },
-        stream(2, pct(40), Some(explicit(200, shares(&[(101, DENOM)])))),
-    ];
-    assert!(validate_weight_schedule(&crossing, 0).is_err());
-
-    let near_end = i64::MAX - 100;
-    let endpoint_overlap = vec![
-        Stream {
-            id: 1,
-            weight: WeightRecord {
-                v_start: pct(49),
-                slope: pct(1) as i64,
-                t_start: near_end,
-                floor: 0,
-                cap: DENOM,
-            },
-            distribution: None,
-        },
-        Stream {
-            id: 2,
-            weight: WeightRecord {
-                v_start: pct(49),
-                slope: pct(1) as i64,
-                t_start: near_end,
-                floor: 0,
-                cap: DENOM,
-            },
-            distribution: Some(explicit(200, shares(&[(101, DENOM)]))),
-        },
-    ];
-    assert!(validate_weight_schedule(&endpoint_overlap, near_end).is_err());
-}
-
-#[test]
-fn solstice_migration_bootstrap_records_have_canonical_anchors() {
-    let activation = 1_000;
-    let slope = (pct(45) / 90) as i64;
-    let consensus = WeightRecord {
-        v_start: pct(95),
-        slope: -slope,
-        t_start: activation,
-        floor: pct(50),
-        cap: pct(95),
-    };
-    let service =
-        WeightRecord { v_start: pct(5), slope, t_start: activation, floor: pct(5), cap: pct(10) };
-
-    validate_weight_record(&consensus).unwrap();
-    validate_weight_record(&service).unwrap();
-    validate_weight_schedule(
-        &[
-            Stream { id: 1, weight: consensus, distribution: None },
-            Stream {
-                id: 2,
-                weight: service,
-                distribution: Some(explicit(200, shares(&[(101, DENOM)]))),
-            },
-        ],
-        activation,
-    )
-    .unwrap();
 }
 
 #[test]
@@ -571,51 +169,6 @@ fn rejects_out_of_band_anchors_on_every_weight_queue_operation() {
         assert_eq!(before, streams);
         assert_eq!(EPOCH_UNDEFINED, next_epoch(&streams));
     }
-}
-
-#[test]
-fn allocates_reward_in_stream_order_and_conserves_attos() {
-    let streams = vec![
-        stream(1, pct(50), None),
-        stream(2, pct(25), Some(explicit(200, shares(&[(101, DENOM)])))),
-        stream(3, pct(10), Some(explicit(201, shares(&[(102, DENOM)])))),
-    ];
-    let reward = TokenAmount::from_atto(7);
-    let allocation = allocate_reward(&streams, 0, &reward).unwrap();
-
-    assert_eq!(TokenAmount::from_atto(3), allocation.miner);
-    assert_eq!(vec![2, 3], allocation.service.iter().map(|row| row.id).collect::<Vec<_>>());
-    assert_eq!(TokenAmount::from_atto(1), allocation.service[0].amount);
-    assert_eq!(TokenAmount::zero(), allocation.service[1].amount);
-    assert_eq!(TokenAmount::from_atto(3), allocation.burn);
-    assert!(allocation.schedule_valid);
-    assert_eq!(
-        reward,
-        &allocation.miner
-            + &allocation.service.iter().fold(TokenAmount::zero(), |sum, row| sum + &row.amount)
-            + &allocation.burn
-    );
-
-    let mut accruals = vec![
-        StreamAccrual { id: 2, amount: TokenAmount::from_atto(4) },
-        StreamAccrual { id: 3, amount: TokenAmount::from_atto(5) },
-    ];
-    accrue_service(&mut accruals, &allocation.service).unwrap();
-    assert_eq!(TokenAmount::from_atto(5), accruals[0].amount);
-    assert_eq!(TokenAmount::from_atto(5), accruals[1].amount);
-
-    let before = accruals.clone();
-    assert!(
-        accrue_service(
-            &mut accruals,
-            &[
-                StreamAccrual { id: 2, amount: TokenAmount::from_atto(1) },
-                StreamAccrual { id: 99, amount: TokenAmount::from_atto(1) },
-            ],
-        )
-        .is_err()
-    );
-    assert_eq!(before, accruals);
 }
 
 #[test]
@@ -687,178 +240,6 @@ fn burns_sentinel_share_without_accruing_or_tombstoning_it() {
         claim(&mut streams, &accruals, 2, &[BURNT_FUNDS_ACTOR_ADDR, Address::new_id(102)]).unwrap();
     assert_eq!(vec![TokenAmount::zero(), TokenAmount::from_atto(5)], claimed);
     assert!(streams.tombstones.is_empty());
-}
-
-#[test]
-fn indivisible_sentinel_portion_preserves_survivor_entitlements() {
-    let third = DENOM / 3;
-    let survivor_shares = shares(&[(101, third), (102, third)]);
-    let reward = TokenAmount::from_atto(2);
-
-    // Control: three ordinary recipients split the full two-atto service pool. Each recipient's
-    // third floors to zero when claimed.
-    let mut ordinary = StreamsState {
-        streams: vec![stream(
-            2,
-            DENOM,
-            Some(explicit(200, shares(&[(101, third), (102, third), (103, DENOM - 2 * third)]))),
-        )],
-        ..Default::default()
-    };
-    let ordinary_allocation = allocate_reward(&ordinary.streams, 0, &reward).unwrap();
-    let mut ordinary_accruals = ordinary_allocation.service.clone();
-    let ordinary_claims =
-        claim(&mut ordinary, &ordinary_accruals, 2, &[Address::new_id(101), Address::new_id(102)])
-            .unwrap();
-    assert_eq!(vec![TokenAmount::zero(), TokenAmount::zero()], ordinary_claims);
-
-    // Sentinel case: the third recipient's share burns. Flooring the survivor pool yields one
-    // atto, so the two surviving claims remain zero. Flooring the burn instead would leave a
-    // two-atto pool and incorrectly pay each survivor one atto after denominator adjustment.
-    let mut sentinel = StreamsState {
-        streams: vec![stream(2, DENOM, Some(explicit(200, survivor_shares)))],
-        ..Default::default()
-    };
-    let sentinel_allocation = allocate_reward(&sentinel.streams, 0, &reward).unwrap();
-    assert_eq!(TokenAmount::from_atto(1), sentinel_allocation.service[0].amount);
-    assert_eq!(TokenAmount::from_atto(1), sentinel_allocation.burn);
-    assert_eq!(
-        ordinary_allocation.service[0].amount,
-        &sentinel_allocation.service[0].amount + &sentinel_allocation.burn
-    );
-
-    let mut sentinel_accruals = sentinel_allocation.service.clone();
-    let sentinel_claims =
-        claim(&mut sentinel, &sentinel_accruals, 2, &[Address::new_id(101), Address::new_id(102)])
-            .unwrap();
-    assert_eq!(ordinary_claims, sentinel_claims);
-
-    accrue_service(
-        &mut ordinary_accruals,
-        &allocate_reward(&ordinary.streams, 0, &reward).unwrap().service,
-    )
-    .unwrap();
-    accrue_service(
-        &mut sentinel_accruals,
-        &allocate_reward(&sentinel.streams, 0, &reward).unwrap().service,
-    )
-    .unwrap();
-    let ordinary_claims =
-        claim(&mut ordinary, &ordinary_accruals, 2, &[Address::new_id(101), Address::new_id(102)])
-            .unwrap();
-    let sentinel_claims =
-        claim(&mut sentinel, &sentinel_accruals, 2, &[Address::new_id(101), Address::new_id(102)])
-            .unwrap();
-    assert_eq!(vec![TokenAmount::from_atto(1), TokenAmount::from_atto(1)], ordinary_claims);
-    assert_eq!(ordinary_claims, sentinel_claims);
-}
-
-#[test]
-fn invalid_weight_envelope_allocates_no_reward_portion() {
-    let streams = vec![
-        stream(1, pct(60), None),
-        stream(2, pct(50), Some(explicit(200, shares(&[(101, DENOM)])))),
-    ];
-    let reward = TokenAmount::from_atto(7);
-
-    let allocation = allocate_reward(&streams, 0, &reward).unwrap();
-
-    assert!(!allocation.schedule_valid);
-    assert_eq!(TokenAmount::zero(), allocation.miner);
-    assert!(allocation.service.is_empty());
-    assert_eq!(TokenAmount::zero(), allocation.burn);
-}
-
-#[test]
-fn malformed_implicit_weight_allocates_no_reward_portion() {
-    let mut malformed = stream(1, 0, None);
-    malformed.weight.cap = DENOM + 1;
-    let streams = vec![malformed, stream(2, pct(20), Some(explicit(200, shares(&[(101, DENOM)]))))];
-    let reward = TokenAmount::from_atto(7);
-
-    let allocation = allocate_reward(&streams, 0, &reward).unwrap();
-
-    assert!(!allocation.schedule_valid);
-    assert_eq!(TokenAmount::zero(), allocation.miner);
-    assert!(allocation.service.is_empty());
-    assert_eq!(TokenAmount::zero(), allocation.burn);
-}
-
-#[test]
-fn malformed_explicit_weight_allocates_no_reward_portion() {
-    let mut malformed = stream(2, 0, Some(explicit(200, shares(&[(101, DENOM)]))));
-    malformed.weight.cap = DENOM + 1;
-    let reward = TokenAmount::from_atto(7);
-
-    let allocation = allocate_reward(&[malformed], 0, &reward).unwrap();
-
-    assert!(!allocation.schedule_valid);
-    assert_eq!(TokenAmount::zero(), allocation.miner);
-    assert!(allocation.service.is_empty());
-    assert_eq!(TokenAmount::zero(), allocation.burn);
-}
-
-#[test]
-fn rejects_invalid_share_maps() {
-    assert!(validate_shares(&shares(&[(101, DENOM - 1)])).is_err());
-    assert!(validate_shares(&shares(&[(101, DENOM / 2), (101, DENOM / 2)])).is_err());
-    assert!(validate_shares(&shares(&[(101, DENOM), (102, 0)])).is_err());
-
-    let too_many: Vec<_> = (0..=MAX_RECIPIENTS)
-        .map(|idx| RecipientShare {
-            recipient: Address::new_id(100 + idx as u64),
-            share: if idx == 0 { DENOM } else { 0 },
-        })
-        .collect();
-    assert!(validate_shares(&too_many).is_err());
-}
-
-#[test]
-fn admits_and_strips_burn_sentinel_rows() {
-    let normalized =
-        normalize_shares(shares(&[(99, pct(20)), (101, pct(50)), (99, pct(30))])).unwrap();
-    assert_eq!(shares(&[(101, pct(50))]), normalized);
-
-    assert!(normalize_shares(shares(&[(99, 0), (101, DENOM)])).is_err());
-    assert!(normalize_shares(shares(&[(99, pct(20)), (101, pct(40)), (101, pct(40)),])).is_err());
-    assert!(normalize_shares(shares(&[(99, DENOM)])).unwrap().is_empty());
-
-    let mut over_limit = shares(&[(101, DENOM - MAX_RECIPIENTS as u64)]);
-    over_limit.extend(
-        (0..MAX_RECIPIENTS).map(|_| RecipientShare { recipient: BURNT_FUNDS_ACTOR_ADDR, share: 1 }),
-    );
-    assert!(normalize_shares(over_limit).is_err());
-}
-
-#[test]
-fn folds_period_under_outgoing_shares_and_burns_only_dust() {
-    let third = DENOM / 3;
-    let old_shares = shares(&[(101, third), (102, third), (103, DENOM - 2 * third)]);
-    let mut distribution = explicit(200, old_shares);
-    distribution.payable.push(RecipientAmount {
-        recipient: Address::new_id(104),
-        amount: TokenAmount::from_atto(5),
-    });
-    distribution.claimed_period.push(RecipientAmount {
-        recipient: Address::new_id(101),
-        amount: TokenAmount::from_atto(1),
-    });
-    let mut streams = StreamsState {
-        streams: vec![stream(2, pct(20), Some(distribution))],
-        ..Default::default()
-    };
-    let mut accruals = vec![StreamAccrual { id: 2, amount: TokenAmount::from_atto(10) }];
-
-    let burn = set_shares(&mut streams, &mut accruals, 2, shares(&[(105, DENOM)])).unwrap();
-    let distribution = streams.streams[0].distribution.as_ref().unwrap();
-    assert_eq!(TokenAmount::from_atto(1), burn);
-    assert_eq!(TokenAmount::zero(), accruals[0].amount);
-    assert!(distribution.claimed_period.is_empty());
-    assert_eq!(shares(&[(105, DENOM)]), distribution.shares);
-    assert_eq!(TokenAmount::from_atto(2), amount(&distribution.payable, 101));
-    assert_eq!(TokenAmount::from_atto(3), amount(&distribution.payable, 102));
-    assert_eq!(TokenAmount::from_atto(3), amount(&distribution.payable, 103));
-    assert_eq!(TokenAmount::from_atto(5), amount(&distribution.payable, 104));
 }
 
 #[test]
@@ -941,174 +322,6 @@ fn caps_payable_rows_at_128_rejects_129_atomically_and_recovers_through_claims()
     assert_eq!(MAX_RECIPIENTS, second_batch.len());
     assert!(second_batch.iter().all(|amount| amount == &TokenAmount::from_atto(1)));
     assert!(streams.streams[0].distribution.as_ref().unwrap().payable.is_empty());
-}
-
-// Wire maps arrive in any order and normalize_shares sorts them; a persisted map that is not
-// ascending by recipient can only have been written by something other than f02.
-#[test]
-fn structural_validation_rejects_unordered_stored_shares() {
-    let ordered = explicit(300, shares(&[(101, pct(50)), (102, pct(50))]));
-    let streams = StreamsState {
-        streams: vec![stream(2, pct(20), Some(ordered.clone()))],
-        ..Default::default()
-    };
-    validate_award_state_structure(&streams).unwrap();
-
-    let mut unordered = ordered;
-    unordered.shares.swap(0, 1);
-    let streams =
-        StreamsState { streams: vec![stream(2, pct(20), Some(unordered))], ..Default::default() };
-    let error = validate_award_state_structure(&streams).unwrap_err();
-    assert_eq!("stored share recipients are not ordered", error.to_string());
-
-    // The same rule gates the initial map carried by a pending registration.
-    let streams = StreamsState {
-        pending_writes: vec![PendingWrite {
-            id: Some(3),
-            op: PendingWriteOp::RegisterStream,
-            payload: RawBytes::serialize(&RegisterStreamPayload {
-                weight: constant_weight(pct(10)),
-                distribution: Some(DistributionInit {
-                    writer: Address::new_id(300),
-                    shares: shares(&[(102, pct(50)), (101, pct(50))]),
-                }),
-            })
-            .unwrap(),
-            effective_epoch: 1,
-        }],
-        ..Default::default()
-    };
-    let error = validate_award_state_structure(&streams).unwrap_err();
-    assert!(error.to_string().contains("stored share recipients are not ordered"), "{error}");
-}
-
-#[test]
-fn structural_validation_rejects_payable_reservation_over_cap() {
-    let mut distribution = explicit(300, full_share_map(100));
-    distribution.payable = (0..=MAX_PAYABLE_ROWS_PER_STREAM)
-        .map(|idx| RecipientAmount {
-            recipient: Address::new_id(100 + idx as u64),
-            amount: TokenAmount::from_atto(1),
-        })
-        .collect();
-    let streams = StreamsState {
-        streams: vec![stream(2, pct(20), Some(distribution))],
-        ..Default::default()
-    };
-
-    let error = validate_award_state_structure(&streams).unwrap_err();
-    assert_eq!(
-        format!(
-            "stream 2 payable row reservation {} exceeds maximum {MAX_PAYABLE_ROWS_PER_STREAM}",
-            MAX_PAYABLE_ROWS_PER_STREAM + 1
-        ),
-        error.to_string()
-    );
-}
-
-#[test]
-fn fold_dust_preserves_the_supply_decomposition_without_moving_counters() {
-    let (mut streams, mut accruals) = base_state();
-    let third = DENOM / 3;
-    let split = shares(&[(101, third), (102, third), (103, DENOM - 2 * third)]);
-    streams.streams[1].distribution.as_mut().unwrap().shares = split.clone();
-    let mut supply = SupplyTracker::default();
-
-    for _ in 0..2 {
-        supply.award(&streams, &mut accruals, 0, TokenAmount::from_atto(51));
-        let before_burn = supply.total_burn.clone();
-        let before_service = supply.total_service.clone();
-        let dust = set_shares(&mut streams, &mut accruals, 2, split.clone()).unwrap();
-        assert_eq!(TokenAmount::from_atto(1), dust);
-        supply.burn_dust(dust);
-        assert_eq!(before_burn, supply.total_burn);
-        assert_eq!(before_service, supply.total_service);
-        supply.assert_invariants(&streams, &accruals);
-    }
-
-    assert_eq!(TokenAmount::from_atto(22), supply.total_burn);
-    assert_eq!(TokenAmount::from_atto(20), supply.total_service);
-    assert_eq!(TokenAmount::from_atto(2), supply.total_dust);
-    supply.assert_invariants(&streams, &accruals);
-}
-
-#[test]
-fn claims_live_and_payable_amounts_once_in_request_order() {
-    let mut distribution = explicit(200, shares(&[(101, DENOM / 2), (102, DENOM - DENOM / 2)]));
-    distribution.payable = vec![
-        RecipientAmount { recipient: Address::new_id(101), amount: TokenAmount::from_atto(3) },
-        RecipientAmount { recipient: Address::new_id(102), amount: TokenAmount::from_atto(4) },
-    ];
-    distribution.claimed_period = vec![RecipientAmount {
-        recipient: Address::new_id(101),
-        amount: TokenAmount::from_atto(2),
-    }];
-    let mut streams = StreamsState {
-        streams: vec![stream(2, pct(20), Some(distribution))],
-        ..Default::default()
-    };
-    let accruals = vec![StreamAccrual { id: 2, amount: TokenAmount::from_atto(11) }];
-    let wallets =
-        [Address::new_id(101), Address::new_id(101), Address::new_id(102), Address::new_id(999)];
-
-    let result = claim(&mut streams, &accruals, 2, &wallets).unwrap();
-    assert_eq!(
-        vec![
-            TokenAmount::from_atto(6),
-            TokenAmount::zero(),
-            TokenAmount::from_atto(9),
-            TokenAmount::zero(),
-        ],
-        result
-    );
-
-    let distribution = streams.streams[0].distribution.as_ref().unwrap();
-    assert!(distribution.payable.is_empty());
-    assert_eq!(TokenAmount::from_atto(5), amount(&distribution.claimed_period, 101));
-    assert_eq!(TokenAmount::from_atto(5), amount(&distribution.claimed_period, 102));
-    let before = streams.clone();
-    let zero = claim(&mut streams, &accruals, 2, &[Address::new_id(999)]).unwrap();
-    assert_eq!(vec![TokenAmount::zero()], zero);
-    assert_eq!(before, streams);
-}
-
-#[test]
-fn claims_tombstones_and_deletes_them_when_drained() {
-    let mut streams = StreamsState {
-        tombstones: vec![Tombstone {
-            id: 3,
-            payable: vec![
-                RecipientAmount {
-                    recipient: Address::new_id(101),
-                    amount: TokenAmount::from_atto(7),
-                },
-                RecipientAmount {
-                    recipient: Address::new_id(102),
-                    amount: TokenAmount::from_atto(8),
-                },
-            ],
-        }],
-        ..Default::default()
-    };
-
-    let result = claim(
-        &mut streams,
-        &[],
-        3,
-        &[Address::new_id(101), Address::new_id(101), Address::new_id(999)],
-    )
-    .unwrap();
-    assert_eq!(vec![TokenAmount::from_atto(7), TokenAmount::zero(), TokenAmount::zero()], result);
-    assert_eq!(1, streams.tombstones.len());
-
-    let result = claim(&mut streams, &[], 3, &[Address::new_id(102)]).unwrap();
-    assert_eq!(vec![TokenAmount::from_atto(8)], result);
-
-    let before = streams.clone();
-    let result =
-        claim(&mut streams, &[], 3, &[Address::new_id(102), Address::new_id(999)]).unwrap();
-    assert_eq!(vec![TokenAmount::zero(), TokenAmount::zero()], result);
-    assert_eq!(before, streams);
 }
 
 #[test]
