+++ actors/reward/src/streams/weights.rs
@@ -1,0 +1,144 @@
+// Copyright 2019-2022 ChainSafe Systems
+// SPDX-License-Identifier: Apache-2.0, MIT
+
+use std::collections::BTreeSet;
+
+use anyhow::{Result, ensure};
+use fvm_ipld_encoding::tuple::*;
+use fvm_shared::clock::ChainEpoch;
+
+use super::{DENOM, Stream, StreamId};
+
+/// A clamped linear weight in `DENOM` fixed point.
+///
+/// Persisted in `Stream` and encoded in weight-management messages and deferred payloads.
+/// Admitted records satisfy `floor <= v_start <= cap <= DENOM`.
+#[derive(Clone, Debug, Default, PartialEq, Eq, Serialize_tuple, Deserialize_tuple)]
+pub struct WeightRecord {
+    pub v_start: u64,
+    pub slope: i64,
+    pub t_start: ChainEpoch,
+    pub floor: u64,
+    pub cap: u64,
+}
+
+/// One stream update in `SetWeightRecords` and `StepWeightRecords` parameters.
+#[derive(Clone, Debug, PartialEq, Eq, Serialize_tuple, Deserialize_tuple)]
+pub struct WeightRecordUpdate {
+    pub id: StreamId,
+    pub weight: WeightRecord,
+}
+
+/// Complete weight-update batch captured in one deferred call.
+#[derive(Clone, Debug, PartialEq, Eq, Serialize_tuple, Deserialize_tuple)]
+pub(super) struct WeightRecordsPayload {
+    pub(super) updates: Vec<WeightRecordUpdate>,
+}
+
+/// Evaluates a weight at `epoch`, clamped to its inclusive floor and cap.
+pub(super) fn compute_weight(record: &WeightRecord, epoch: ChainEpoch) -> u64 {
+    let delta = i128::from(epoch) - i128::from(record.t_start);
+    // |delta| <= 2^64 - 1 and |slope| <= 2^63, so the product fits i128.
+    let product = i128::from(record.slope) * delta;
+    // Saturation affects only malformed v_start > DENOM and is equivalent before the u64 clamp.
+    let value = i128::from(record.v_start).saturating_add(product);
+    u64::try_from(value.min(i128::from(record.cap)).max(i128::from(record.floor)))
+        .expect("bounded weight fits u64")
+}
+
+pub(super) fn validate_weight_record(record: &WeightRecord) -> Result<()> {
+    ensure!(record.floor <= record.cap, "weight floor exceeds cap");
+    ensure!(record.v_start >= record.floor, "weight v_start is below floor");
+    ensure!(record.v_start <= record.cap, "weight v_start exceeds cap");
+    ensure!(record.cap <= DENOM, "weight cap exceeds DENOM");
+    Ok(())
+}
+
+pub(super) fn validate_weight_updates(updates: &[WeightRecordUpdate]) -> Result<()> {
+    ensure!(!updates.is_empty(), "weight-record update is empty");
+    for pair in updates.windows(2) {
+        ensure!(pair[0].id != pair[1].id, "duplicate weight-record stream ID {}", pair[0].id);
+        ensure!(pair[0].id <= pair[1].id, "weight-record updates are not ordered");
+    }
+    for update in updates {
+        validate_weight_record(&update.weight)?;
+    }
+    Ok(())
+}
+
+/// Returns epochs at or after `start_epoch` that can change an admitted record's regime.
+/// Adjacent epochs bracket integer crossings so validation cannot skip a one-epoch violation.
+pub(super) fn weight_breakpoints(
+    record: &WeightRecord,
+    start_epoch: ChainEpoch,
+) -> Vec<ChainEpoch> {
+    // Include the validation boundary and any later record anchor.
+    let mut epochs = BTreeSet::from([start_epoch]);
+    if record.t_start >= start_epoch {
+        epochs.insert(record.t_start);
+    }
+
+    if record.slope != 0 {
+        // A canonical anchor has one crossing in the slope's forward direction. The opposite
+        // crossing can matter only when validation begins before the anchor.
+        let start = i128::from(record.t_start);
+        let value = i128::from(record.v_start);
+        let slope = i128::from(record.slope);
+        // The numerator and quotient are within +/-u64::MAX, leaving ample i128 headroom.
+        let mut insert_crossing = |bound: u64| {
+            let quotient = (i128::from(bound) - value) / slope;
+            for offset in [-1_i128, 0, 1] {
+                let epoch = start + quotient + offset;
+                if let Ok(epoch) = ChainEpoch::try_from(epoch)
+                    && epoch >= start_epoch
+                {
+                    epochs.insert(epoch);
+                }
+            }
+        };
+
+        insert_crossing(if record.slope > 0 { record.cap } else { record.floor });
+        if start_epoch < record.t_start {
+            insert_crossing(if record.slope > 0 { record.floor } else { record.cap });
+        }
+    }
+
+    // Sample beyond the last crossing and at the epoch domain's absolute endpoint.
+    if let Some(last) = epochs.last().copied()
+        && let Some(after) = last.checked_add(1)
+    {
+        epochs.insert(after);
+    }
+    epochs.insert(ChainEpoch::MAX);
+    epochs.into_iter().collect()
+}
+
+/// Validates aggregate stream weight from `start_epoch` onward.
+pub(super) fn validate_weight_schedule(streams: &[Stream], start_epoch: ChainEpoch) -> Result<()> {
+    validate_weight_schedule_through(streams, start_epoch, ChainEpoch::MAX)
+}
+
+/// Validates aggregate stream weight over the inclusive epoch interval.
+fn validate_weight_schedule_through(
+    streams: &[Stream],
+    start_epoch: ChainEpoch,
+    end_epoch: ChainEpoch,
+) -> Result<()> {
+    ensure!(end_epoch >= start_epoch, "weight schedule interval is reversed");
+    let mut epochs = BTreeSet::from([start_epoch, end_epoch]);
+    for stream in streams {
+        validate_weight_record(&stream.weight)?;
+        epochs.extend(
+            weight_breakpoints(&stream.weight, start_epoch)
+                .into_iter()
+                .take_while(|epoch| *epoch <= end_epoch),
+        );
+    }
+
+    for epoch in epochs {
+        let sum: u128 =
+            streams.iter().map(|stream| u128::from(compute_weight(&stream.weight, epoch))).sum();
+        ensure!(sum <= u128::from(DENOM), "stream weights exceed DENOM at epoch {epoch}: {sum}");
+    }
+    Ok(())
+}
